// src/lib/utils/enums.ts
export enum SidebarMode {
  EXPANDED = "expanded",
  COLLAPSED = "collapsed",
}

export enum Theme {
  DARK = "dark",
  LIGHT = "light",
  SYSTEM = "system",
}

export enum ViewportSize {
  MOBILE = "mobile",
  TABLET = "tablet",
  DESKTOP = "desktop",
}

export enum MenuItemType {
  SIMPLE = "simple",
  SECTION = "section",
  DIVIDER = "divider",
}

export enum UserRole {
  ADMIN = "ADMIN",
  USER = "USER",
  SPAM_SELL_ONLY = "SPAM_SELL_ONLY",
}

export enum Exchange {
  BINANCE = "binance",
  BITGET = "bitget",
  BYBIT = "bybit",
  GATEIO = "gate.io",
  INDODAX = "indodax",
  KUCOIN = "kucoin",
  MEXC = "mexc",
  OKX = "okx",
}

export enum AlertType {
  Info = "info",
  Success = "success",
  Warning = "warning",
  Error = "error",
}

export enum ViewMode {
  TABLE = "table",
  CARDS = "cards",
}